src/services/log-collector/requirements.txt
watchdog
pyyaml
Flask